backend/rag_ingest/requirements.txt
fastapi
InstructorEmbedding
langchain==0.0.267
chromadb==0.3.29
sentence-transformers
docx2txt
unstructured
openpyxl
PyYAML==6.0.1
uvicorn==0.23.2
Werkzeug==2.3.6
utils
slowapi
pymongo
cachetools
logger
markdown